Add tabs, move radio activity contents into fragment
