+++ app/build.gradle
@@ -46,5 +46,4 @@ dependencies {
     compile "com.android.support:appcompat-v7:$supportLibVersion"
     compile "com.android.support:support-v4:$supportLibVersion"
     compile "com.android.support:design:$supportLibVersion"
-    compile 'com.android.support.constraint:constraint-layout:1.0.2'
 }
