+++ app/src/main/res/layout/fragment_radio.xml
@@ -1,0 +1,115 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="fill_parent"
+    android:layout_height="fill_parent"
+    android:orientation="vertical"
+    android:fitsSystemWindows="true"
+    android:padding="16dp"
+    tools:context=".ui.activities.MainActivity">
+
+    <ImageView
+        android:id="@+id/logo"
+        android:layout_width="326dp"
+        android:layout_height="133dp"
+        android:layout_gravity="center_horizontal"
+        android:contentDescription="@string/infoVolume"
+        app:srcCompat="@drawable/main_logo" />
+
+    <TextView
+        android:id="@+id/nowPlaying"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:fontFamily="sans-serif"
+        android:text="@string/nowPlaying"
+        android:textAlignment="center"
+        android:textColor="@android:color/white"
+        android:textSize="18sp" />
+
+    <TextView
+        android:id="@+id/requestedText"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="@string/requestedText"
+        android:textAlignment="center"
+        android:textColor="@android:color/white"
+        android:textColorLink="@color/colorAccent"
+        android:textSize="14sp" />
+
+    <LinearLayout
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center_horizontal">
+
+        <ImageView
+            android:id="@+id/volumeDown"
+            android:layout_width="32dp"
+            android:layout_height="32dp"
+            android:contentDescription="@string/infoVolumeDown"
+            android:scaleType="fitXY"
+            android:tint="@android:color/white"
+            app:srcCompat="@drawable/ic_volume_down_black_24dp" />
+
+        <SeekBar
+            android:id="@+id/seekBar"
+            android:layout_width="250dp"
+            android:layout_height="wrap_content"
+            android:max="100"
+            android:progress="50"
+            android:progressBackgroundTint="@color/volumeLight"
+            android:progressTint="@color/colorAccent"
+            android:thumbTint="@color/colorAccent" />
+
+        <ImageView
+            android:id="@+id/volumeUp"
+            android:layout_width="32dp"
+            android:layout_height="32dp"
+            android:contentDescription="@string/infoVolumeUp"
+            android:scaleType="fitXY"
+            android:tint="@android:color/white"
+            app:srcCompat="@drawable/ic_volume_up_black_24dp" />
+
+    </LinearLayout>
+
+    <LinearLayout
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center_horizontal">
+
+        <ImageButton
+            android:id="@+id/playPause"
+            android:layout_width="48dp"
+            android:layout_height="48dp"
+            android:background="@android:color/transparent"
+            android:backgroundTint="@android:color/transparent"
+            android:contentDescription="@string/infoPlayPause"
+            android:foreground="@android:color/transparent"
+            android:foregroundTint="@android:color/transparent"
+            android:scaleType="fitXY"
+            android:tint="@android:color/white"
+            android:visibility="visible"
+            app:srcCompat="@drawable/ic_play_arrow_black_24dp" />
+
+        <ImageButton
+            android:id="@+id/favoriteButton"
+            android:layout_width="48dp"
+            android:layout_height="48dp"
+            android:background="@android:color/transparent"
+            android:backgroundTint="@android:color/transparent"
+            android:scaleType="fitXY"
+            android:tint="@android:color/white"
+            app:srcCompat="@drawable/ic_star_border_black_24dp"
+            tools:ignore="ContentDescription" />
+
+    </LinearLayout>
+
+    <TextView
+        android:id="@+id/currentText"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="@string/currentListeners"
+        android:textAlignment="center"
+        android:textColor="@android:color/white" />
+
+</LinearLayout>
